refactor: inline pure-derived aboutPanel and topbarBadge stores
Both stores held no own state, only a computed deriving badges from
extensionStore. Each had a single consumer, so the derivation moves into
the consuming component and the stores are deleted.

Left useActionBarButtonStore and useHelpCenterStore (shared across 2+
consumers) and useNodeHelpStore (dedicated test) as stores.
