+++ src/components/topbar/TopbarBadges.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="flex h-full shrink-0 items-center">
     <TopbarBadge
-      v-for="badge in topbarBadgeStore.badges"
+      v-for="badge in badges"
       :key="badge.text"
       :badge
       :display-mode="displayMode"
@@ -15,7 +15,8 @@
 import { breakpointsTailwind, useBreakpoints } from '@vueuse/core'
 import { computed } from 'vue'
 
-import { useTopbarBadgeStore } from '@/stores/topbarBadgeStore'
+import { useExtensionStore } from '@/stores/extensionStore'
+import type { TopbarBadge as TopbarBadgeType } from '@/types/comfy'
 
 import TopbarBadge from './TopbarBadge.vue'
 
@@ -34,5 +35,8 @@ const displayMode = computed<'full' | 'compact' | 'icon-only'>(() => {
   return 'icon-only'
 })
 
-const topbarBadgeStore = useTopbarBadgeStore()
+const extensionStore = useExtensionStore()
+const badges = computed<TopbarBadgeType[]>(() =>
+  extensionStore.extensions.flatMap((e) => e.topbarBadges ?? [])
+)
 </script>
