+++ src/stores/README.md
@@ -108,7 +108,6 @@ The following table lists ALL 46 store instances in the system as of 2026-01-29:
 
 | File                         | Store                         | Description                                             | Category   |
 | ---------------------------- | ----------------------------- | ------------------------------------------------------- | ---------- |
-| aboutPanelStore.ts           | useAboutPanelStore            | Manages the About panel state and badges                | UI         |
 | apiKeyAuthStore.ts           | useApiKeyAuthStore            | Handles API key authentication                          | Auth       |
 | comfyManagerStore.ts         | useComfyManagerStore          | Manages ComfyUI application state                       | Core       |
 | comfyRegistryStore.ts        | useComfyRegistryStore         | Handles extensions registry                             | Registry   |
